+++ lib/src/flutter_taglib_bindings.dart
@@ -143,6 +143,10 @@ ffi.Pointer<ffi.Char> taglib_bridge_get_format(
   return generated.taglib_bridge_get_format(file);
 }
 
+int taglib_bridge_is_lossless(ffi.Pointer<TagLibBridgeFile> file) {
+  return generated.taglib_bridge_is_lossless(file);
+}
+
 int taglib_bridge_has_cover(ffi.Pointer<TagLibBridgeFile> file) {
   return generated.taglib_bridge_has_cover(file);
 }